simulator: emulate accelerometer (99% implemented)
There's a gravity_vector in the simulator that points in the direction
of gravity, and there's a little pictograph of the badge that you can
rotate in 3D with the mouse to change the simulated orientation of the
badge.

To complete the implementation, we just need to hook the accelerometer
API to this gravity vector.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>